feat(contracts:br): shrink grid on piece move
- loops through all pieces and shrinks grid
- basic test for functionality

+++ contracts/src/common/BRGame.sol
@@ -10,7 +10,7 @@ enum BRGameStatus {
 struct BRGame {
   uint256 startTime;
   uint32 rechargeTime;
-  uint32 initialGridDim;
-  uint32 secondsPerGridShrink;
+  uint16 initialGridDim;
+  uint16 secondsPerGridShrink;
   BRGameStatus status;
 }
